requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==3.2
django-celery-beat==2.5.0
django-filter==23.5
django-timezone-field==6.1.0
djangorestframework==3.14.0
Faker==21.0.0
flower==2.0.1
humanize==4.9.0
kombu==5.3.4
Markdown==3.5.1
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
